+++ client/js/components/text-editor.js
