+++ client/js/index.js
@@ -4,7 +4,7 @@ console.log(`Client running in ${process.env.NODE_ENV} mode`);
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import MainContainer from './components/main-container'
+import PanelContainer from './components/panel-container'
 import store from './store';
 import {Provider} from 'react-redux';
 
@@ -12,7 +12,7 @@ import {Provider} from 'react-redux';
 document.addEventListener('DOMContentLoaded', () => {
 	ReactDOM.render(
 		<Provider store={store}>
-			<MainContainer />
+			<PanelContainer />
 		</Provider>, 
 		document.getElementById('app'))});
